+++ lib/libesp32/berry/src/be_code.h
@@ -13,7 +13,7 @@
 #define be_code_freeregs(f, n)      ((f)->freereg -= (bbyte)(n))
 
 int be_code_allocregs(bfuncinfo *finfo, int count);
-void be_code_prebinop(bfuncinfo *finfo, int op, bexpdesc *e);
+void be_code_prebinop(bfuncinfo *finfo, int op, bexpdesc *e, int dst);
 void be_code_binop(bfuncinfo *finfo, int op, bexpdesc *e1, bexpdesc *e2, int dst);
 int be_code_unop(bfuncinfo *finfo, int op, bexpdesc *e);
 int be_code_setvar(bfuncinfo *finfo, bexpdesc *e1, bexpdesc *e2, bbool keep_reg);